Move Flipper files on Android
