+++ android/app/src/release/java/io/metamask/ReactNativeFlipper.java
@@ -4,9 +4,11 @@
  * <p>This source code is licensed under the MIT license found in the LICENSE file in the root
  * directory of this source tree.
  */
-package com.rndiffapp;
+package io.metamask;
+
 import android.content.Context;
 import com.facebook.react.ReactInstanceManager;
+
 /**
  * Class responsible of loading Flipper inside your React Native application. This is the release
  * flavor of it so it's empty as we don't want to load Flipper.
